add rest controller for ingredient

+++ src/main/java/com/example/pizzeria/dto/ingredient/IngredientResponse.java
@@ -3,6 +3,6 @@ package com.example.pizzeria.dto.ingredient;
 public record IngredientResponse(
         Long id,
         String name,
-        int price
+        double price
 ) {
 }
