+++ src/main/java/com/example/pizzeria/model/Product.java
@@ -2,14 +2,14 @@ package com.example.pizzeria.model;
 
 public abstract class Product {
     private String name;
-    private int price;
+    private double price;
 
-    public Product(String name, int price){
+    public Product(String name, double price){
         this.name = name;
         this.price = price;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 
@@ -21,7 +21,7 @@ public abstract class Product {
         this.name = name;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
     }
